Adding files from 19/02
More progress made but a lot of time was spent mulling over abstract classes and inheritance. Trying to implement the fields and the different particles, considering how they should be in different classes, but it's what they should be inheriting... Especially with what should be an abstract method, since the point is to limit the number of times that we re-write code.