dimensions of fields
EM fields now both have dimensions that can be specified. If dimensions are not specified,
the field is considered to cover all space. If the dimensions are specified, the field can be constrained
to only affect a particle that exist within its dimensions, which are just a cuboid right now.