randomised particle creation
The pickled file is a saved file for 10 interacting protons, no external fields.
the particles are all made by the bunch, needs a lot of care to make sure that the energy of 
the particles is appropriate. Probably best to precalculate the initial velocity that you want before
you input the mean energy. likewise for the bunch spread.