Saving and plotting improved
Saving of data is improved slightly, changes made to the particlebunchclass and how it returns data,
lots of new plots for the graphs and this also includes an upload of a pickle of the constrained data.
accelerator is working pretty much perfectly tbh. Don't know what else to add as features now.