Updates to Run sim and plotting
Lots of small tweaks to the simulation specifics, but mainly just that plotting is now handled by
the Plotting file and PlottingClass. Considerations have been made to add extra complications to 
the plotting but ignored. 
Consider a final plot of the effect of altering the phase of the accelerating electric field, with no
external confining electric fields. Although the confining electric fields do look very nice.
Finally to do: move the running and saving of simulation state to its own file