phase plotting and simulation classes
big big overhauls.
simulations are now made into their own classes, and an abstract class for simulations is also included to assist with building new simulation classes.
the standard simulation is overhauled so that it functions within a class.
the new simulation that is 50-100 simulations and saving the final std dev. in energy from all of them and plotting that against a 
phase change is also included.
data files from running 4 protons for under a single revolution are included. these best show the reduction in spread a phase shift can have.
Simulations still run from main.py